+++ api-request/documentation/routes/home-element/template.js
@@ -1,0 +1,28 @@
+import md from '../../../README.md';
+
+export const template = (context, html) => html`
+<main-layout>
+<main-header slot="page-header">
+  <header>
+    <h1 class="content-width">${context.__('@aofl/api-request')}</h1>
+  </header>
+</main-header>
+
+<main slot="page-content" class="content-width column">
+  <tab-headers id="content-tabs" groupId="contentTabs" storeInstance$="${context.storeInstance}">
+    <aofl-list-option slot="tab" class="link" value="overview" selected$="${context.selectedContentTab === 'overview'}">Overview</aofl-list-option>
+    <aofl-list-option slot="tab" class="link" value="examples" selected$="${context.selectedContentTab === 'examples'}">Examples</aofl-list-option>
+  </tab-headers>
+
+  <tab-content groupId="contentTabs" tabId="overview" storeInstance$="${context.storeInstance}">${html([md])}</tab-content>
+  <tab-content groupId="contentTabs" tabId="examples" storeInstance$="${context.storeInstance}">
+    <aofl-preview dom-scope="api-request-preview">
+      <span slot="title">apiRequest</span>
+      <code slot="javascript">${context.jsExample}</code>
+      <code slot="css">${context.cssExample}</code>
+      <code slot="htmlmixed">${context.htmlExample}</code>
+    </aofl-preview>
+  </tab-content>
+</main>
+</main-layout>
+`;