+++ api-request/src/formatter-manager/index.js
@@ -2,7 +2,7 @@
  * formatter manager hold reference to available formatter layouts
  *
  * @summary FormatterManager
- * @version 1.0.0-beta.1
+ * @version 1.0.0
  * @author Isaac Yoon <isaac.yoon@aofl.com>
  * @author Arian Khosravi <arian.khosravi@aofl.com>
  *
@@ -21,8 +21,10 @@ class FormatterManager {
    * @param {String} format name of the format
    * @return {Object}
    */
-  getFormmatter(format) {
-    return this.formatters[format];
+  getFormatter(format) {
+    if (this.formatters.hasOwnProperty(format)) {
+      return this.formatters[format];
+    }
   }
 
   /**
@@ -30,8 +32,18 @@ class FormatterManager {
    *
    * @param {String} format name of the format
    * @param {Object} formatter formatter object
+   * @throws {TypeError} Will throw an error when the provided pormatter doesn't implement pack() and unpack()
+   * @throws {Error} Will throw an error when attempting to replace an existing formatter
    */
   addFormatter(format, formatter) {
+    if (typeof formatter.pack !== 'function' || typeof formatter.unpack !== 'function') {
+      throw new TypeError('formatter must implement pack() and unpack() functions');
+    }
+
+    if (typeof this.formatters[format] !== 'undefined') {
+      throw new Error(`${format} already exists. Cannot replace an existing format.`);
+    }
+
     this.formatters[format] = formatter;
   }
 }