+++ polyfill-service/src/object-assign-polyfill/index.js
@@ -1,3 +1,4 @@
+/* istanbul ignore next */
 /* eslint-disable */
 if (typeof Object.assign != 'function') {
   // Must be writable: true, enumerable: false, configurable: true